dev: rename accrual pkg to connection
